chore: move send calling message out of repository into handler
