+++ logic/src/androidMain/kotlin/com/wire/kalium/logic/feature/call/GlobalCallManager.kt
@@ -13,6 +13,7 @@ import com.wire.kalium.logic.data.call.CallRepository
 import com.wire.kalium.logic.data.client.ClientRepository
 import com.wire.kalium.logic.data.id.QualifiedID
 import com.wire.kalium.logic.data.user.UserRepository
+import com.wire.kalium.logic.feature.message.MessageSender
 import com.waz.log.LogHandler as NativeLogHandler
 
 actual class GlobalCallManager(
@@ -44,14 +45,16 @@ actual class GlobalCallManager(
         callRepository: CallRepository,
         userRepository: UserRepository,
         clientRepository: ClientRepository,
-        callMapper: CallMapper
+        callMapper: CallMapper,
+        messageSender: MessageSender
     ): CallManager {
         return callManagerHolder[userId] ?: CallManagerImpl(
             calling = calling,
             callRepository = callRepository,
             userRepository = userRepository,
             clientRepository = clientRepository,
-            callMapper = callMapper
+            callMapper = callMapper,
+            messageSender = messageSender
         ).also {
             callManagerHolder[userId] = it
         }
