+++ logic/src/androidMain/kotlin/com/wire/kalium/logic/feature/call/scenario/OnHttpRequest.kt
@@ -1,22 +1,27 @@
 package com.wire.kalium.logic.feature.call.scenario
 
+import com.benasher44.uuid.uuid4
 import com.sun.jna.Pointer
 import com.wire.kalium.calling.Calling
 import com.wire.kalium.calling.types.Handle
-import com.wire.kalium.logic.data.call.CallRepository
+import com.wire.kalium.logic.CoreFailure
 import com.wire.kalium.logic.data.conversation.ClientId
 import com.wire.kalium.logic.data.id.ConversationId
+import com.wire.kalium.logic.data.message.Message
+import com.wire.kalium.logic.data.message.MessageContent
 import com.wire.kalium.logic.data.user.UserId
+import com.wire.kalium.logic.feature.message.MessageSender
 import com.wire.kalium.logic.functional.Either
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Deferred
 import kotlinx.coroutines.launch
+import kotlinx.datetime.Clock
 
 //TODO create unit test
 class OnHttpRequest(
     private val handle: Deferred<Handle>,
     private val calling: Calling,
-    private val callRepository: CallRepository,
+    private val messageSender: MessageSender,
     private val callingScope: CoroutineScope
 ) {
     fun sendHandlerSuccess(
@@ -28,7 +33,7 @@ class OnHttpRequest(
     ) {
         callingScope.launch {
             messageString?.let { message ->
-                when (callRepository.sendCallingMessage(conversationId, avsSelfUserId, avsSelfClientId, message)) {
+                when (sendCallingMessage(conversationId, avsSelfUserId, avsSelfClientId, message)) {
                     is Either.Right -> {
                         calling.wcall_resp(
                             inst = handle.await(),
@@ -48,5 +53,17 @@ class OnHttpRequest(
                 }
             }
         }
+    }
+
+    private suspend fun sendCallingMessage(
+        conversationId: ConversationId,
+        userId: UserId,
+        clientId: ClientId,
+        data: String
+    ): Either<CoreFailure, Unit> {
+        val messageContent = MessageContent.Calling(data)
+        val date = Clock.System.now().toString()
+        val message = Message(uuid4().toString(), messageContent, conversationId, date, userId, clientId, Message.Status.SENT)
+        return messageSender.sendMessage(message)
     }
 }
