+++ logic/src/androidMain/kotlin/com/wire/kalium/logic/feature/call/scenario/OnSendOTR.kt
@@ -6,12 +6,12 @@ import com.wire.kalium.calling.callbacks.SendHandler
 import com.wire.kalium.calling.types.Handle
 import com.wire.kalium.calling.types.Size_t
 import com.wire.kalium.logic.callingLogger
-import com.wire.kalium.logic.data.call.CallRepository
 import com.wire.kalium.logic.data.conversation.ClientId
 import com.wire.kalium.logic.data.id.toConversationId
 import com.wire.kalium.logic.data.user.toUserId
 import com.wire.kalium.logic.feature.call.AvsCallBackError
 import com.wire.kalium.logic.feature.call.CallManagerImpl
+import com.wire.kalium.logic.feature.message.MessageSender
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Deferred
 
@@ -21,14 +21,14 @@ class OnSendOTR(
     private val calling: Calling,
     private val selfUserId: String,
     private val selfClientId: String,
-    private val callRepository: CallRepository,
+    private val messageSender: MessageSender,
     private val callingScope: CoroutineScope
 ) : SendHandler {
     override fun onSend(
         context: Pointer?,
         conversationId: String,
-        avsSelfUserId: String,
-        avsSelfClientId: String,
+        userIdSelf: String,
+        clientIdSelf: String,
         userIdDestination: String?,
         clientIdDestination: String?,
         data: Pointer?,
@@ -36,17 +36,17 @@ class OnSendOTR(
         isTransient: Boolean,
         arg: Pointer?
     ): Int {
-        return if (selfUserId != avsSelfUserId && selfClientId != avsSelfClientId) {
+        return if (selfUserId != userIdSelf && selfClientId != clientIdSelf) {
             callingLogger.i("${CallManagerImpl.TAG} -> sendHandler error")
             AvsCallBackError.INVALID_ARGUMENT.value
         } else {
             callingLogger.i("${CallManagerImpl.TAG} -> sendHandler success")
-            OnHttpRequest(handle, calling, callRepository, callingScope).sendHandlerSuccess(
+            OnHttpRequest(handle, calling, messageSender, callingScope).sendHandlerSuccess(
                 context = context,
                 messageString = data?.getString(0, CallManagerImpl.UTF8_ENCODING),
                 conversationId = conversationId.toConversationId(),
-                avsSelfUserId = avsSelfUserId.toUserId(),
-                avsSelfClientId = ClientId(avsSelfClientId)
+                avsSelfUserId = userIdSelf.toUserId(),
+                avsSelfClientId = ClientId(clientIdSelf)
             )
             AvsCallBackError.NONE.value
         }
