+++ logic/src/commonMain/kotlin/com/wire/kalium/logic/data/call/CallRepository.kt
@@ -3,14 +3,8 @@ package com.wire.kalium.logic.data.call
 import com.benasher44.uuid.uuid4
 import com.wire.kalium.logic.CoreFailure
 import com.wire.kalium.logic.callingLogger
-import com.wire.kalium.logic.data.conversation.ClientId
-import com.wire.kalium.logic.data.id.ConversationId
-import com.wire.kalium.logic.data.message.Message
-import com.wire.kalium.logic.data.message.MessageContent
-import com.wire.kalium.logic.data.user.UserId
 import com.wire.kalium.logic.feature.call.Call
 import com.wire.kalium.logic.feature.call.CallStatus
-import com.wire.kalium.logic.feature.message.MessageSender
 import com.wire.kalium.logic.functional.Either
 import com.wire.kalium.logic.functional.suspending
 import com.wire.kalium.logic.wrapApiRequest
@@ -20,7 +14,6 @@ import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.map
-import kotlinx.datetime.Clock
 import kotlin.math.max
 
 interface CallRepository {
@@ -32,17 +25,10 @@ interface CallRepository {
     fun createCall(call: Call)
     fun updateCallStatusById(conversationId: String, status: CallStatus)
     fun updateCallParticipants(conversationId: String, participants: List<Participant>)
-    suspend fun sendCallingMessage(
-        conversationId: ConversationId,
-        userId: UserId,
-        clientId: ClientId,
-        data: String
-    ): Either<CoreFailure, Unit>
 }
 
 internal class CallDataSource(
-    private val callApi: CallApi,
-    private val messageSender: MessageSender
+    private val callApi: CallApi
 ) : CallRepository {
 
     //TODO to be saved somewhere ?
@@ -73,18 +59,6 @@ internal class CallDataSource(
 
     override fun getOngoingCall(): Flow<List<Call>> = allCalls.map {
         it.calls.values.filter { call -> call.status == CallStatus.ESTABLISHED }
-    }
-
-    override suspend fun sendCallingMessage(
-        conversationId: ConversationId,
-        userId: UserId,
-        clientId: ClientId,
-        data: String
-    ): Either<CoreFailure, Unit> {
-        val messageContent = MessageContent.Calling(data)
-        val date = Clock.System.now().toString()
-        val message = Message(uuid4().toString(), messageContent, conversationId, date, userId, clientId, Message.Status.SENT)
-        return messageSender.sendMessage(message)
     }
 
     override fun createCall(call: Call) {
