+++ logic/src/commonTest/kotlin/com/wire/kalium/logic/data/call/CallRepositoryTest.kt
@@ -1,6 +1,5 @@
 package com.wire.kalium.logic.data.call
 
-import com.wire.kalium.logic.feature.message.MessageSender
 import com.wire.kalium.logic.util.shouldSucceed
 import com.wire.kalium.network.api.call.CallApi
 import com.wire.kalium.network.utils.NetworkResponse
@@ -19,16 +18,12 @@ class CallRepositoryTest {
     @Mock
     private val callApi = mock(classOf<CallApi>())
 
-    @Mock
-    private val messageSender = mock(classOf<MessageSender>())
-
     private lateinit var callRepository: CallRepository
 
     @BeforeTest
     fun setUp() {
         callRepository = CallDataSource(
-            callApi = callApi,
-            messageSender
+            callApi = callApi
         )
     }
