+++ logic/src/jvmMain/kotlin/com/wire/kalium/logic/feature/call/GlobalCallManager.kt
@@ -5,6 +5,7 @@ import com.wire.kalium.logic.data.call.CallRepository
 import com.wire.kalium.logic.data.client.ClientRepository
 import com.wire.kalium.logic.data.id.QualifiedID
 import com.wire.kalium.logic.data.user.UserRepository
+import com.wire.kalium.logic.feature.message.MessageSender
 
 actual class GlobalCallManager {
 
@@ -13,6 +14,7 @@ actual class GlobalCallManager {
         callRepository: CallRepository,
         userRepository: UserRepository,
         clientRepository: ClientRepository,
-        callMapper: CallMapper
+        callMapper: CallMapper,
+        messageSender: MessageSender
     ): CallManager = CallManagerImpl()
 }
